WIP: Update MRI projects for Aluminium
This bumps versions to the following:
- odlparent-7.0.1
- yangtools-5.0.1
- mdsal-6.0.0
- controller-2.0.0

Implements the List->Map migration due to the MD-SAL 6 upgrade and
partially the migration to Uint{8,16,32} types.

WIP: So far, only lisp-proto is migrated, which likely needed the most
changes.

+++ mappingservice/lisp-proto/src/main/java/org/opendaylight/lispflowmapping/lisp/serializer/address/InstanceIdSerializer.java
@@ -16,6 +16,7 @@ import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.lisp.addres
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.lisp.address.types.rev151105.LispAddress;
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.lisp.address.types.rev151105.lisp.address.address.InstanceId;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.lfm.lisp.proto.rev151105.eid.container.Eid;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 public final class InstanceIdSerializer extends LcafSerializer {
     private static final int MAX_INSTANCE_ID = 16777216;
@@ -69,7 +70,7 @@ public final class InstanceIdSerializer extends LcafSerializer {
         if (instanceId > MAX_INSTANCE_ID) {
             throw new LispSerializationException("Instance ID is longer than 24 bits: " + instanceId);
         }
-        ctx.setVni(new InstanceIdType(instanceId));
+        ctx.setVni(new InstanceIdType(Uint32.valueOf(instanceId)));
         return LispAddressSerializer.getInstance().deserializeEid(buffer, ctx);
     }
